+++ demuxer.py
@@ -29,7 +29,6 @@ class Demuxer:
         self.blacklist = b              # VCID blacklist
         self.keys = k                   # Decryption keys
         self.channelHandlers = {}       # List of channel handlers
-        self.vcduCounter = -1           # VCDU continuity counter
 
         if self.downlink == "LRIT":
             self.coreWait = 54          # Core loop delay in ms for LRIT (108.8ms per packet @ 64 kbps)
@@ -78,9 +77,6 @@ class Demuxer:
                     if self.verbose: print("SPACECRAFT \"{}\" NOT SUPPORTED".format(vcdu.SCID))
                     continue
 
-                # Check VCDU continuity counter
-                self.continuity(vcdu)
-
                 # Check for VCID change
                 if lastVCID != vcdu.VCID:
                     if self.verbose: print()
@@ -115,27 +111,6 @@ class Demuxer:
                 dumpFile.close()
             
             return
-
-    def continuity(self, vcdu):
-        """
-        Checks VCDU packet continuity by comparing packet counters
-        """
-
-        # If at least one VCDU has been received
-        if self.vcduCounter != -1:
-            # Check counter reset
-            if self.vcduCounter == 16777215 and vcdu.COUNTER == 0:
-                self.vcduCounter = vcdu.COUNTER
-                return
-            
-            diff = vcdu.COUNTER - self.vcduCounter - 1
-            if diff != 0:
-                if self.verbose:
-                    print("  DROPPED {} PACKETS    (CURRENT: {}   LAST: {}   VCID: {})".format(diff, vcdu.COUNTER, self.vcduCounter, vcdu.VCID))
-                else:
-                    print("  DROPPED {} PACKETS".format(diff))
-        
-        self.vcduCounter = vcdu.COUNTER
 
     def push(self, packet):
         """
@@ -191,6 +166,7 @@ class Channel:
         """
 
         self.VCID = vcid            # VCID for this handler
+        self.counter = -1           # VCDU continuity counter
         self.verbose = v            # Verbose output flag
         self.crclut = crclut        # CP_PDU CRC LUT
         self.outputPath = output    # xRIT file output path root
@@ -204,6 +180,9 @@ class Channel:
         Takes in VCDUs for the channel handler to process
         :param packet: Parsed VCDU object
         """
+
+        # Check VCDU continuity counter
+        self.continuity(vcdu)
 
         # Parse M_PDU
         mpdu = CCSDS.M_PDU(vcdu.MPDU)
@@ -263,8 +242,30 @@ class Channel:
                 self.cCPPDU.append(mpdu.PACKET)
             except AttributeError:
                 if self.verbose: print("  NO CP_PDU TO APPEND M_PDU TO (DROPPED PACKETS?)")
-
     
+
+    def continuity(self, vcdu):
+        """
+        Checks VCDU packet continuity by comparing packet counters
+        """
+
+        # If at least one VCDU has been received
+        if self.counter != -1:
+            # Check counter reset
+            if self.counter == 16777215 and vcdu.COUNTER == 0:
+                self.counter = vcdu.COUNTER
+                return
+            
+            diff = vcdu.COUNTER - self.counter - 1
+            if diff > 0:
+                if self.verbose:
+                    print("  DROPPED {} PACKETS    (CURRENT: {}   LAST: {}   VCID: {})".format(diff, vcdu.COUNTER, self.counter, vcdu.VCID))
+                else:
+                    print("  DROPPED {} PACKETS".format(diff))
+        
+        self.counter = vcdu.COUNTER
+    
+
     def check_CPPDU(self, lenok, crcok):
         """
         Checks length and CRC of finished CP_PDU