+++ CHANGELOG.md
@@ -13,7 +13,7 @@ All notable changes to this project will be documented in this file.
   - 
 
 ### Fixed
-  - 
+  - VCDU continuity counter
 </details>